// src/app/Models/GetSelectBaseClass.ts
// Generated by https://quicktype.io

export class GetSelectBaseClass {
    EmpID:      string;
    EmpCode:    string;
    EmpNameC:   string;
    EmpNameE:   string;
    JobID:      number;
    JobName:    string;
    ChiefCode:  string;
    EffectDate: string;
    DeptaID:    number;
    PosType:    string;
    DeptaName:  string;
}